feat: split custom fields into separate components for edit permission
